git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/release2.06@7263 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/gamesys/Event.cpp
@@ -67,7 +67,7 @@ void idEventDef::Construct()
 	assert(name != NULL);
 	assert(!idEvent::initialized);
 	
-	for (size_t i = 0; i < args.size(); i++)
+	for ( int i = 0; i < args.size(); i++ )
 		formatspec[i] = args[i].type;
 	formatspec[args.size()] = '\0';
